move ShepherdState methods and refs to new util class.
