+++ src/main/java/org/ecocean/shepherd/core/ShepherdPMF.java
@@ -25,9 +25,6 @@ public class ShepherdPMF {
     private static TreeMap<String, PersistenceManagerFactory> pmfs = new TreeMap<String,
         PersistenceManagerFactory>();
 
-    private static ConcurrentHashMap<String, String> shepherds = new ConcurrentHashMap<String,
-        String>();
-
     public synchronized static PersistenceManagerFactory getPMF(String context) {
         // public static PersistenceManagerFactory getPMF(String dbLocation) {
         if (pmfs == null) { pmfs = new TreeMap<String, PersistenceManagerFactory>(); }
@@ -406,25 +403,5 @@ public class ShepherdPMF {
             }
         }
         return myProps;
-    }
-
-    public static void setShepherdState(String shepherdID, String state) {
-        if (shepherds == null) shepherds = new ConcurrentHashMap<String, String>();
-        shepherds.put(shepherdID, state);
-    }
-
-    public static void removeShepherdState(String shepherdID) {
-        if (shepherds == null) shepherds = new ConcurrentHashMap<String, String>();
-        shepherds.remove(shepherdID);
-    }
-
-    public static String getShepherdState(String shepherdID) {
-        if (shepherds == null) shepherds = new ConcurrentHashMap<String, String>();
-        return shepherds.get(shepherdID);
-    }
-
-    public static ConcurrentHashMap<String, String> getAllShepherdStates() {
-        if (shepherds == null) shepherds = new ConcurrentHashMap<String, String>();
-        return shepherds;
     }
 }
